move the test implementations into the tests folder
