+++ tests/unittests/Infrastructure/NoopDatabase.php
@@ -1,6 +1,8 @@
 <?php
 
-namespace Gdatacyberdefenseag\GdataAntivirus\Infrastructure\Database;
+namespace unittests\Infrastructure;
+
+use Gdatacyberdefenseag\GdataAntivirus\Infrastructure\Database\IGdataAntivirusDatabase;
 
 class NoopDatabase implements IGdataAntivirusDatabase {
     public function query( $query, ...$args ): int|bool {
